requirements.txt
torch
torchvision
torchaudio
transformers>=4.41.0
accelerate
datasets
evaluate
jiwer
Pillow
regex
editdistance
sentencepiece
timm
